backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
pytest==8.2.1
httpx==0.27.0
python-jose==3.3.0
auth0-python==4.7.1
alembic==1.13.1
openai==1.35.10
python-dotenv==1.0.1
pgvector==0.3.0
tiktoken==0.7.0